Move tests + fix name generation tests
