+++ decompiler/util/default.json
@@ -286,24 +286,6 @@
                 "argument_name": "--empty-for-loops"
             },
             {
-                "dest": "readability-based-refinement.for_loop_variable_names",
-                "default": [
-                    "i",
-                    "j",
-                    "k",
-                    "l",
-                    "m",
-                    "n"
-                ],
-                "type": "array",
-                "elementType": "string",
-                "title": "Rename for-loop variables into desired names",
-                "description": "Rename for-loop variables to values from list",
-                "is_hidden_from_gui": false,
-                "is_hidden_from_cli": false,
-                "argument_name": "--for-loop-variable-names"
-            },
-            {
                 "dest": "readability-based-refinement.max_condition_complexity_for_loop_recovery",
                 "default": 100,
                 "type": "number",
@@ -348,16 +330,6 @@
                 "argument_name": "--for-loop-exclude-conditions"
             },
             {
-                "dest": "readability-based-refinement.rename_while_loop_variables",
-                "default": true,
-                "type": "boolean",
-                "title": "Rename while-loop variables",
-                "description": "Rename while-loop counter variables to counter, counter1, ...",
-                "is_hidden_from_gui": false,
-                "is_hidden_from_cli": false,
-                "argument_name": "--rename-while-loop-variables"
-            },
-            {
                 "dest": "variable-name-generation.notation",
                 "default": "system_hungarian",
                 "title": "Variable Naming Convention",
@@ -371,6 +343,34 @@
                 "is_hidden_from_gui": false,
                 "is_hidden_from_cli": false,
                 "argument_name": "--variable-generation-notation"
+            },
+            {
+                "dest": "variable-name-generation.rename_while_loop_variables",
+                "default": true,
+                "type": "boolean",
+                "title": "Rename while-loop variables",
+                "description": "Rename while-loop counter variables to counter, counter1, ...",
+                "is_hidden_from_gui": false,
+                "is_hidden_from_cli": false,
+                "argument_name": "--rename-while-loop-variables"
+            },
+            {
+                "dest": "variable-name-generation.for_loop_variable_names",
+                "default": [
+                    "i",
+                    "j",
+                    "k",
+                    "l",
+                    "m",
+                    "n"
+                ],
+                "type": "array",
+                "elementType": "string",
+                "title": "Rename for-loop variables",
+                "description": "Rename for-loop variables to values from given list",
+                "is_hidden_from_gui": false,
+                "is_hidden_from_cli": false,
+                "argument_name": "--for-loop-variable-names"
             },
             {
                 "dest": "variable-name-generation.variable_name",
